+++ src/lib/storeMeta.ts
@@ -1,7 +1,14 @@
 import fs from 'fs'
-import MIMETypes from './MIMETypes'
 import IResponse from '../interfaces/IResponse'
 import IResponseMeta from '../interfaces/IResponseMeta'
+import {
+  isBreakingSelfDescriptiveness,
+  isForgettingHypermedia,
+  isIgnoringCaching,
+  isIgnoringMIMEType,
+  isIgnoringStatusCode,
+  isMisusingCookies,
+} from './designAntipatternsDetectors'
 
 export const storeResponseMeta = async (
   uri: string,
@@ -12,9 +19,9 @@ export const storeResponseMeta = async (
 
   const bodyObject: object = JSON.parse(bodyText)
 
-  // console.log(
-  //   isForgettingHypermedia(res, bodyObject, httpMethod)
-  // )
+  console.log(
+    isForgettingHypermedia(res, bodyObject, httpMethod)
+  )
 
   const responseMeta: IResponseMeta = {
     uri,
@@ -48,113 +55,6 @@ export const storeResponseMeta = async (
   writeToFile(responseMeta)
 
   console.log(`Stored info for ${uri}`)
-}
-
-// TODO put checks in separate file
-
-// TODO unit test all of this
-
-function isBreakingSelfDescriptiveness(
-  res: IResponse,
-  httpMethod: String
-) {
-  // TODO: ignore Etag here, covered in ignoring caching check
-  // for now based on this: https://www.oreilly.com/library/view/rest-api-design/9781449317904/ch04.html
-
-  const encouragedHeaders = [
-    'Content-Type',
-    'Content-Length',
-  ]
-
-  if (httpMethod.toUpperCase() === 'GET') {
-    encouragedHeaders.push('Last-Modified')
-  }
-
-  if (
-    httpMethod.toUpperCase() === 'POST' ||
-    httpMethod.toUpperCase() === 'PUT' ||
-    httpMethod.toUpperCase() === 'PATCH'
-  ) {
-    encouragedHeaders.push('Location')
-  }
-
-  for (const header of encouragedHeaders) {
-    if (!res.headers.has(header)) return true
-  }
-
-  return false
-}
-
-function isForgettingHypermedia(
-  res: IResponse,
-  body: object,
-  httpMethod: string
-) {
-  const result = getAllProperties(body)
-  console.log(result)
-}
-
-/**
- * Inspired by method described here:
- * https://stackoverflow.com/a/11922384/9374593
- * @param obj
- */
-function getAllProperties(obj: any): any | void {
-  const properties = []
-
-  for (const property in obj) {
-    const value = obj[property]
-
-    typeof value === 'object'
-      ? properties.push(getAllProperties(value))
-      : // TODO here, check if property is correct property
-        // i.e. link/links/href, if so can set flag bool
-        properties.push(property)
-  }
-
-  return properties
-}
-
-function isIgnoringCaching(
-  res: IResponse,
-  httpMethod: string
-): boolean {
-  const cacheControlElements = res.headers
-    .get('Cache-Control')
-    ?.split(', ')
-
-  return (
-    httpMethod.toUpperCase() !== 'GET' || // Only checks for ignoring caching antipattern in GET requests
-    !res.headers.has('Etag') ||
-    !res.headers.has('Cache-Control') ||
-    (cacheControlElements !== undefined &&
-      cacheControlElements?.includes('no-cache')) ||
-    (cacheControlElements !== undefined &&
-      cacheControlElements?.includes('no-store'))
-  )
-}
-
-function isIgnoringMIMEType(res: IResponse) {
-  return !MIMETypes.some((type) =>
-    res.headers.get('content-type')?.includes(type)
-  )
-}
-
-function isIgnoringStatusCode(
-  res: IResponse,
-  httpMethod: string
-) {
-  // TODO perhaps check this more thoroughly, check for acceptable status code for various http methods
-  return (
-    httpMethod.toUpperCase() !== 'GET' && res.status === 200
-  )
-}
-
-function isMisusingCookies(res: IResponse) {
-  return (
-    res.headers.has('set-cookie') ||
-    res.headers.has('cookie')
-  )
 }
 
 function writeToFile(responseMeta: IResponseMeta) {
